Duplicate registration_ids for GCM Devices no longer allowed
There are some unresolved database limitations related to
the maximum allowed size of a GCM registration_id that
are causing client issues. This commit addresses the
issue for clients by applying validation at the API layer
for now. However, it is still possible to enter a duplicate
registration_id using through the admin interface
or shell_plus.

See also: #184, #197, #216